requirements.txt
streamlit
google-generativeai
python-dotenv
langchain-community
langchain-text-splitters
PyPDF2
faiss-cpu
sentence-transformers
huggingface-hub